requirements-test.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-test.in
#
attrs==22.2.0
    # via pytest
black==22.3.0
    # via -r requirements-test.in
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via black
coverage[toml]==7.1.0
    # via pytest-cov
exceptiongroup==1.1.0
    # via pytest
factory-boy==3.2.1
    # via
    #   -r requirements-test.in
    #   pytest-factoryboy
faker==16.7.0
    # via factory-boy
flake8==3.9.1
    # via -r requirements-test.in
freezegun==1.2.2
    # via -r requirements-test.in
idna==3.4
    # via requests
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==2.0.0
    # via pytest
isort==5.8.0
    # via -r requirements-test.in
mccabe==0.6.1
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==23.0
    # via pytest
pathspec==0.11.0
    # via black
platformdirs==3.0.0
    # via black
pluggy==1.0.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pytest==7.2.1
    # via
    #   -r requirements-test.in
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
pytest-cov==4.0.0
    # via -r requirements-test.in
pytest-django==4.5.2
    # via -r requirements-test.in
pytest-factoryboy==2.5.1
    # via -r requirements-test.in
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
requests==2.28.2
    # via requests-mock
requests-mock==1.10.0
    # via -r requirements-test.in
six==1.16.0
    # via
    #   python-dateutil
    #   requests-mock
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   pytest
typing-extensions==4.4.0
    # via
    #   black
    #   pytest-factoryboy
urllib3==1.26.14
    # via requests
